+++ lib/runners/processor/golangci_lint.rb
@@ -75,12 +75,20 @@ module Runners
     def analyzer_options
       [].tap do |opts|
         analysis_targets.each { |target| opts << target }
+
+        # NOTE: Always override the value in `.golangci.yml`
         opts << "--out-format=json"
         opts << "--print-issued-lines"
         opts << "--print-linter-name"
         opts << "--issues-exit-code=0"
         opts << "--color=never"
         opts << "--concurrency=2"
+
+        # NOTE: The value should be less than the top-level timeout value.
+        #
+        # @see https://github.com/sider/runners/blob/1231de0ca047c7a23449ab1a7bb0751f39f16643/images/Dockerfile.end.erb#L11
+        opts << "--timeout=15m"
+
         opts << "--tests=#{config_linter[:tests]}" unless config_linter[:tests].nil?
         opts << "--config=#{path_to_config}" if path_to_config
         Array(config_linter[:disable]).each { |disable| opts << "--disable=#{disable}" }
@@ -100,7 +108,7 @@ module Runners
     def path_to_config
       return config_linter[:config] if config_linter[:config]
 
-      # @see https://github.com/golangci/golangci-lint#config-file
+      # @see https://golangci-lint.run/usage/configuration/#config-file
       default_config_file_is_found = %w[.golangci.yml .golangci.toml .golangci.json].find { |f| (current_dir / f).exist? }
       return if default_config_file_is_found